# O(n^2) -> O(n) implementation for scale_free_graph