@florishermsen 2021-04-07
Almost @rossbar Better formulated would be that we approach uniform sampling when `len(nodes) * delta` >> `len(degree_distribution)` where `degree_distribution` (sampling `candidates` in the code) are the "raveled" in and out-degree distributions (i.e. a node with degree 7 appears 7 times in that list), so its length is equal to `len(nodes) * avg(degree)`. As you can see the inequality then simplifies to `delta` >> `avg(degree)`.

This should make sense because going back to the algorithm, the probability for a node to be sampled scales with its in or out-degree plus the bias, i.e. `~(d+b)`. In case `b` >> `d`, this effectively becomes `~b`, and since the bias is the same for all nodes this indeed effectively is uniform sampling.