@rossbar 2021-04-10
Hmm so looking at this again, it looks to me like there may have been a very subtle bug that would persist in this implementation as well. For example, look at the following (original) code chunk:

https://github.com/networkx/networkx/blob/f97705493ac84f99aa6ab5aae3129d5dc2578e2c/networkx/generators/directed.py#L276-L279

The comment indicates that `v` should be a node that is not already present in the Graph. It seems this relies on assumptions about the nodes in G; most likely based on the default value of `create_using`, which creates a graph starting with node 0 and incrementing upwards by one. However, it's easy to violate this assumption e.g.

```python
>>> G = nx.MultiDiGraph([(2, 3), (3, 4), (2, 4)])
>>> H = nx.scale_free_graph(10, create_using=G)
```

In this case, if `scale_free_graph` hit the snippet above, the "new" node would be `3`, which is clearly already in the Graph. I'm not sure what impact this has on the algorithm as a whole (maybe this is fine?) - it's just a detail where the implementation seems to diverge from the desired behavior (gathered from the comments) for certain graph inputs. Is this a problem for the algorithm? If it's not, at the very least the comments should then be updated to reflect what can actually happen in the general case.