@florishermsen 2021-04-10
@rossbar I added a commit (https://github.com/networkx/networkx/pull/4727/commits/9bc7d31e80c7379d91cb5499c4beb2e63d279db7) that addresses precisely this!

I explained it further in one of the threads above since I responded to some of the content of your comment there but I guess you missed that, it's tricky to keep these conversations linear..!

-----------------
_Copy of https://github.com/networkx/networkx/pull/4727#discussion_r608503785_ :

Great catch! I've added a commit that properly supports working with existing graphs by introducing a node state (simple list of nodes) and a cursor for new nodes that starts at the highest already occupied `int` + 1.

The code would be simpler if we would sample from `list(G.nodes())` but it would make the code orders of magnitude slower once again for large `n`. The current approach does not sacrifice any of the performance gains, timing results are still similar to posted above.

@rossbar your snippet now produces
```python
NodeView((1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 513, 514, 515, 516, 517, 518, 519, 520, 521, 522))
```

But it now also supports e.g. strings:
```python
>>> G = nx.scale_free_graph(10)
>>> H = nx.relabel_nodes(G, {i: str(2**i) for i in range(10)})
>>> H.nodes()
NodeView(('1', '2', '4', '8', '16', '32', '64', '128', '256', '512'))
>>> K = nx.scale_free_graph(20, create_using=H, delta_in=1e6, delta_out=1e6)
>>> K.nodes()
NodeView(('1', '2', '4', '8', '16', '32', '64', '128', '256', '512', 0, 1, 2, 3, 4, 5, 6, 7, 8, 9))
```