@dschult 2021-04-10
That's nice approach to finding new nodes. But it is fragile in the sense that 1.0 equates to 1 but isn't an integer.
So a graph with nodes 0.0 1.0 2.0 3.0 would have `cursor` be 0 even though it should be 4.

We could use `isinstance(node, numbers.Number)` in place of `type(i)==int`.  Or even `type(node) in (int, float, complex)`. But we might have to keep adding types to that list...
However we do it, we should make sure that counter is an integer -- perhaps when we add the 1: `counter = int(max(...)) + 1`.