@florishermsen 2021-04-10
@dschult interesting, at first I did not understand why in your example the cursor cannot be zero, as I figured the graph would allow for `int 0` and `float 0.0` to be different nodes. Apparently it doesn't, as per normal python dictionary behavior (didn't know that was the implementation)

I'll prepare a fix to properly determine the starting point for the cursor. Treatment that would work for complex as well could be selecting the max + 1 from `int(n.real)` for all numbers n.

Open to suggestions.