@dschult 2021-04-11
Hmmm....   This is getting annoying and obtuse. :}
Perhaps the best way is to implement an easy -- imperfect method for finding new nodes and then let the user override it with parameter e.g. `new_nodes`. We could force `new_nodes` to be a list of `n` nodes. Or we could be fancy and let it be either a starting value for new nodes or a list of new nodes. But the easy/simple interface might look something like:

```python
if new_nodes is None:
    ...current code to find max...
    cursor = itertools.count(max + 1)
else:
    try:
        assert len(new_nodes) == n:
        cursor = iter(new_nodes)
    # catch objects with no `len` or `iter` capability
    except (TypeError, AssertionError):
        raise nx.NetworkXError("new_nodes should be a list of `n` node objects")
```
By making `cursor` an iterator we can use/update it in one step with `next(cursor)`